Create QuickFingerControls
This is for Maya.
This is just a quick way of creating some basic controls for fingers in an automated way using the naming of your joints. I typically make two groups above the controls, one to hold the transforms, and the other to hold drivers for driven keys, or multi parents for different space switches.