Create LocatorAtCenterOfVerts
Probably the handiest little snippet I use all the time. It allows for placing a locator at the center of a selection of verts. Makes quick work of placing joints on mechanical things or just good for laying down a baseline. Obviously not the most complicated things, but I've used it so much and it got me more into finding, modifying, and writing python bits to help myself out.