ParentJointsToControls
This one taught me something important yesterday. I was having such a time with how lists make Unicode names, and trying to always format them became a real pain. Found out, that in a 'for loop' it pulls it as a string which is so stinking convenient and made this logic so much simpler. Might need to go back an tidy up my formatting lines, and will def take this moving forward.